+++ src/locales/default.js
@@ -1,6 +1,8 @@
 export const messages = {
   generators: {
+    elder: 'Elder Mods',
     incursion: 'Incursion',
+    shaped: 'Shaped Mods',
     showcase: 'all'
   }
 };